refactor(hindsight): extract monitor JSONL writing into a module
Move the comparison-record building and the slim quote/route/transaction
projection out of monitor.rs into resolve::jsonl, leaving monitor to drive the
solver and delegate output. No behaviour change.
