+++ tools/hindsight/src/resolve/mod.rs
@@ -7,6 +7,7 @@
 //! and back-of-block (N).
 
 mod compare;
+mod jsonl;
 pub(crate) mod monitor;
 
 use alloy::primitives::{Address, U256};
